test fixes for windows
